requirements.txt
# Core Dependencies - Always Required
autogen-agentchat>=0.4.0
autogen-ext>=0.4.0
openai>=1.0.0
tiktoken>=0.7.0

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
websockets>=11.0.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Machine Learning and Feedback Engine
scikit-learn>=1.3.0
joblib>=1.3.0

# Content Processing
markdown>=3.5.0
python-dotenv>=1.0.0

# Security and Encryption
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4

# Database Support
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Development and Testing (commented out for production)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# Optional Enhanced Features (uncomment as needed)
# plotly>=5.15.0              # For advanced dashboard visualizations
# torch>=2.0.0                # For advanced ML routing
# transformers>=4.30.0        # For semantic analysis
# sentence-transformers>=2.2.0 # For embedding-based routing